refactor: `@blocksuite/editor` -> `@blocksuite/presets`
