+++ .commitlintrc.json
@@ -4,7 +4,16 @@
     "scope-enum": [
       2,
       "always",
-      ["page", "edgeless", "database", "virgo", "store", "std", "playground"]
+      [
+        "page",
+        "edgeless",
+        "database",
+        "virgo",
+        "store",
+        "std",
+        "presets",
+        "playground"
+      ]
     ]
   }
 }
