+++ package.json
@@ -6,7 +6,7 @@
     "packages/block-std",
     "packages/blocks",
     "packages/docs",
-    "packages/editor",
+    "packages/presets",
     "packages/global",
     "packages/lit",
     "packages/playground",
@@ -33,11 +33,11 @@
     "test:headed": "playwright test --debug",
     "test:store": "pnpm --filter @blocksuite/store test",
     "test:virgo": "pnpm --filter @blocksuite/virgo test",
-    "test:integration": "pnpm --filter @blocksuite/editor test",
+    "test:integration": "pnpm --filter @blocksuite/presets test",
     "test:node": "node --experimental-specifier-resolution=node --loader ts-node/esm ./scripts/node-import-test.ts",
     "format": "prettier --write --cache packages tests",
-    "build": "pnpm build:editor && pnpm build:playground && pnpm build:docs",
-    "build:editor": "pnpm --filter @blocksuite/editor build",
+    "build": "pnpm build:presets && pnpm build:playground && pnpm build:docs",
+    "build:presets": "pnpm --filter @blocksuite/presets build",
     "build:playground": "pnpm --filter @blocksuite/playground build",
     "build:docs": "pnpm --filter @blocksuite/docs build",
     "size:data": "node --loader ts-node/esm ./scripts/size-data.ts",
