+++ packages/blocks/src/_presets/index.ts
@@ -19,7 +19,7 @@ import { DataViewBlockSchema } from '../data-view-block/index.js';
 import { DatabaseBlockSchema } from '../database-block/database-model.js';
 import { DatabaseService } from '../database-block/database-service.js';
 import { DividerBlockSchema } from '../divider-block/divider-model.js';
-import { embedGithubBlockSpec } from '../embed-github-block/index.js';
+import { EmbedGithubBlockSpec } from '../embed-github-block/index.js';
 import { FrameBlockSchema } from '../frame-block/index.js';
 import { ImageBlockSchema } from '../image-block/image-model.js';
 import { ImageService } from '../image-block/index.js';
@@ -190,7 +190,7 @@ export const PagePreset: BlockSpec[] = [
       },
     },
   },
-  embedGithubBlockSpec,
+  EmbedGithubBlockSpec,
 ];
 
 export const EdgelessPreset: BlockSpec[] = [
@@ -281,5 +281,5 @@ export const EdgelessPreset: BlockSpec[] = [
       component: literal`affine-surface-ref`,
     },
   },
-  embedGithubBlockSpec,
+  EmbedGithubBlockSpec,
 ];
