+++ packages/blocks/src/embed-github-block/embed-github-spec.ts
@@ -3,7 +3,7 @@ import { literal } from 'lit/static-html.js';
 import { createEmbedBlock } from '../_common/embed-block-helper/index.js';
 import { EmbedGithubBlockModel } from './embed-github-model.js';
 
-export const embedGithubBlockSpec = createEmbedBlock({
+export const EmbedGithubBlockSpec = createEmbedBlock({
   schema: {
     name: 'github',
     version: 1,
