+++ packages/playground/apps/starter/main.ts
@@ -1,9 +1,9 @@
 /// <reference types="./env.d.ts" />
 import '@blocksuite/blocks';
-import '@blocksuite/editor';
+import '@blocksuite/presets';
 import './components/start-panel';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import '@blocksuite/editor/themes/affine.css';
+import '@blocksuite/presets/themes/affine.css';
 
 import { TestUtils } from '@blocksuite/blocks';
 import { ContentParser } from '@blocksuite/blocks/content-parser';
