+++ packages/playground/examples/basic/index.html
@@ -9,8 +9,8 @@
   <body>
     <simple-affine-editor></simple-affine-editor>
     <script type="module">
-      import '@blocksuite/editor';
-      import '@blocksuite/editor/themes/affine.css';
+      import '@blocksuite/presets';
+      import '@blocksuite/presets/themes/affine.css';
     </script>
   </body>
 </html>
