+++ scripts/size-generate.sh
@@ -3,7 +3,7 @@
 set -eu -o pipefail
 
 rm -fr ./temp/size
-pnpm run build:editor
+pnpm run build:presets
 pnpm run size:data
 pnpm -C ./packages/playground run build:size
 pnpm run --silent size:report >./size-report.md
