+++ tests/utils/declare-test-window.ts
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/no-restricted-imports */
 import type { ContentParser } from '../../packages/blocks/src/content-parser.js';
 import type { TestUtils } from '../../packages/blocks/src/index.js';
-import type { EditorContainer } from '../../packages/editor/src/index.js';
 import type { BlockSuiteRoot } from '../../packages/lit/src/index.js';
 import type { DebugMenu } from '../../packages/playground/apps/starter/components/debug-menu.js';
+import type { EditorContainer } from '../../packages/presets/src/index.js';
 import type {
   BaseBlockModel,
   Page,
@@ -20,7 +20,7 @@ declare global {
       global: {
         utils: typeof import('../../packages/global/src/utils.js');
       };
-      editor: typeof import('../../packages/editor/src/index.js');
+      editor: typeof import('../../packages/presets/src/index.js');
     };
     workspace: Workspace;
     ContentParser: typeof ContentParser;
